requirements.txt
Flask==2.2.5
Flask-CORS==3.0.10
mysql-connector-python==8.0.32
python-dotenv==0.21.1
Werkzeug==2.2.3

# NLP and ML with spaCy
spacy==3.5.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
numpy==1.23.5
scikit-learn==1.2.2
python-dateutil==2.8.2
joblib==1.2.0

# Utilities
tqdm==4.65.0
requests==2.28.2
